interface1/requirements.txt
numpy==1.24.4
Flask
gunicorn
easyocr
psycopg2-binary
pdf2image>=1.16.0
PyPDF2>=2.0.0
pillow
opencv-python-headless
face_recognition
fuzzywuzzy
python-Levenshtein
dlib
configparser
poppler-utils
